# Improve Lua UFunction call performance