Profiled 3k ufunction calls in a loop:
* 12800ms spent.
* 99% of that spent in FindObjects, fixed via GetFunctionByNameInChain
* 40ms spent allocating wstrings, fixed via introducing `to_const_wstring` pool
* 25ms spent in FindProperty, fixed via https://github.com/Re-UE4SS/UEPseudo/pull/44/commits/6e14bc4b0b5b0b08378df458609f4fd0e8d0b370

end result is 3k ufunction calls in <9ms on my machine

Testing performed was mostly running a mod on my current target that uses a variety of functionality, though I'm probably only accessing ~20 UFunctions and ~50 properties on 7 different classes of UObject.